+++ ie/c-tag.go
@@ -18,7 +18,7 @@ func NewCTAG(flags, pcp, deiFlag uint8, cvid uint16) *IE {
 		return nil
 	}
 
-	return New(CTAG, b)
+	return New(uint16(CTAG), b)
 }
 
 // CTAG returns CTAG in structured format if the type of IE matches.
