+++ ie/ie_uint16_test.go
@@ -134,8 +134,8 @@ func TestUint16IEs(t *testing.T) {
 			decoderFunc: func(i *ie.IE) (uint16, error) { return i.NumberOfReports() },
 		}, {
 			description: "OffendingIE",
-			structured:  ie.NewOffendingIE(ie.Cause),
-			decoded:     ie.Cause,
+			structured:  ie.NewOffendingIE(uint16(ie.Cause)),
+			decoded:     uint16(ie.Cause),
 			decoderFunc: func(i *ie.IE) (uint16, error) { return i.OffendingIE() },
 		}, {
 			description: "PDRID",
