+++ ie/remote-gtp-u-peer.go
@@ -18,7 +18,7 @@ func NewRemoteGTPUPeer(flags uint8, v4, v6 string, di uint8, ni string) *IE {
 		return nil
 	}
 
-	return New(RemoteGTPUPeer, b)
+	return New(uint16(RemoteGTPUPeer), b)
 }
 
 // RemoteGTPUPeer returns RemoteGTPUPeer in *RemoteGTPUPeerFields if the type of IE matches.
